pmake - rename rmake to pmake now that refactor is complete
